escape swizzle-like parameter names
some shaders have parameters named "bg" or "rb". those cause issues
because mpv puts those parameters in the shader using defines like so:

    const _1234 = 42.69;
    #define bg _1234

where _1234 is a random identifier, and "bg" is the actual parameter
name. so glsl syntax errors arise when the shader tries to do stuff like

    vec3 my_color = vec3(1,2,3);
    vec2 some_value = my_color.bg;

which is then replaced by that by the preprocessor:

    vec3 my_color = vec3(1,2,3);
    vec2 some_value = my_color._1234;